# Conflicts:
#	lib/service/services/signal_r/model/recurring_buys_model.dart
#	lib/shared/constants.dart
#	lib/shared/providers/service_providers.dart

+++ lib/shared/services/remote_config_service/remote_config_values.dart
@@ -18,9 +18,9 @@ late int minAmountOfCharsInPassword;
 late int maxAmountOfCharsInPassword;
 late int quoteRetryInterval;
 late String defaultAssetIcon;
-late int emailResendCountdown;
-late int withdrawalConfirmResendCountdown;
-late int localPinLength;
+late double emailResendCountdown;
+late double withdrawalConfirmResendCountdown;
+late double localPinLength;
 
 // Versioning
 late String recommendedVersion;
